+++ examples/website/src/pages/docs/project_structure.rs
@@ -27,7 +27,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-resuma = { version = "0.2", default-features = false }
+resuma = { version = "0.3", default-features = false }
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1", features = ["derive"] }"#)}